+++ CHANGELOG.md
@@ -11,6 +11,15 @@
 - OpenTelemetry logs exporter now exports logs in OpenTelemetry Log Data Model format instead of formatted strings.
 - (pro) Improved advanced smart crop.
 
+## [4.0.6] - TBA
+### Changed
+- Remove `form` tags from SVGs during sanitization.
+- Use more strict `Content-Security-Policy` header: `default-src 'none'; style-src 'unsafe-inline'; img-src data:; sandbox`.
+
+### Fixed
+- Don't write the `Cache-Control` header for error responses.
+- Allow zero `IMGPROXY_MAX_REDIRECTS` config value.
+
 ## [4.0.5] - 2026-06-12
 ### Fixed
 - Fix handling of wide-gamut ICC profiles.